refactor(web): step 9 — extract chat-fork-dialog and chat-user-input-form from chat-pane
